Reorganized the CM test so it should pass on py2.4.
